run cophe.susie using susie to detect separate signals

Check if a variant causally associated in one trait might be causal in another trait

cophe.susie(
  dataset,
  querysnpid,
  querytrait,
  p1 = 1e-04,
  p2 = 1e-04,
  p12 = 1e-05,
  pa = NULL,
  pc = NULL,
  susie.args = list()
)

Arguments

dataset

either a list with specifically named elements defining the dataset to be analysed. (see check_dataset), or the result of running runsusie on such a dataset

querysnpid

Id of the query variant

querytrait

Query trait name

p1

prior probability a SNP is associated with trait 1, default 1e-4 (coloc prior)

p2

prior probability a SNP is associated with trait 2, default 1e-4 (coloc prior)

p12

prior probability a SNP is associated with both traits, default 1e-5 (coloc prior)

pa

prior probability that a non-query variant is causally associated with the query trait , default \(pa = p2\) (cophescan prior)

pc

prior probability that the query variant is causally associated with the query trait, default \(pc = p12/p1+p12\) (cophescan prior)

susie.args

a named list of additional arguments to be passed to runsusie

Value

a list, containing elements

summary a data.table of posterior probabilities of each global hypothesis, one row per pairwise comparison of signals from the two traits

results a data.table of detailed results giving the posterior probability for each snp to be jointly causal for both traits assuming Hc is true. Please ignore this column if the corresponding posterior support for H4 is not high.

priors a vector of the priors used for the analysis

Author

Ichcha Manipur

Examples

library(cophescan)
data(cophe_multi_trait_data)
query_trait_1 <- cophe_multi_trait_data$summ_stat[['Trait_1']]
querysnpid <- cophe_multi_trait_data$querysnpid
query_trait_1$LD <- cophe_multi_trait_data$LD
res.susie <- cophe.susie(query_trait_1, querysnpid = querysnpid, querytrait='Trait_1')
#> running max iterations: 100
#> 	converged: TRUE
#> SNP Priors
#> 0.8092909090909091e-040.0909090909090909
#> Hypothesis Priors
#> 0.8092909090909090.09980.0909090909090909
#> Running cophe.susie...
#> Using 1000 and 999 available
#> 5.56e-050.03060.969
#> PP for causal query variant: 96.9%
summary(res.susie)
#>   nsnps           hit1           hit2        PP.Hn      PP.Ha     PP.Hc
#> 1  1000 chr19-11182353 chr19-11182144 5.558347e-05 0.03061269 0.9693317
#>     lBF.Ha   lBF.Hc       querysnp  typeBF idx1 idx2 querytrait
#> 1 15.31003 11.95277 chr19-11182353 susieBF    1    1    Trait_1